When homing, the LS don't mean done
Depdending on the homing sequence, the motor may drive against a limit switch,
return and search for the home switch.

When the limit switch is hit, motorRecord assumes "done",
even if the sequence is not completet.

Whenever homing, don't use (pmr->rhls || pmr->rlls) to synthesize "done",
but wait until the controller returns "done".