motorRecord.cc: Fix for SPMG=Pause and CALLBACK_DATA
The following sequence did not work:
- Set SPMG to Pause
- Change the VAL field to a new value
- The poller reports a new position (and moto not moving)

The motor not moving does not mean "motor has stopped".
When SPMG is "Pause", it means, "motor is paused".

Handle ths situation more gracefully:
Don't go into the state-machine at all.

Note: This may need a bigger rework:
Writing to a "drive field" causes .DMOV to "blink",
regardless if the old value is the same or different as the
new value.
How should the interact with SPMG==STOP or SPMG==PAUSE ?