asynMotorController: Fix for wrong timeout in pollAll()
The method "asynStatus xyzcontroller::poll()" returns an asynStatus.
What should happen with the status is not well defined:
The upstream motor ignores it.
Here we add a 1 second sleep, when status != asynSuccess is returned.
However, there is a problem here:
When the pasynUserController_ is not used (and NULL) and
poll() returns asynError during the start of the system,
then asynStatusConnected_ is never reset to asynSuccess and
the poller uses always a 1 second timeout.

This bug was reported by the ecmc users.

Solution:
- Set asynStatusConnected_ as the result of controller::poll()
  (This is new)
- Set it to asynError when axis::poll() fails
  (This is old)
- Use the fast poll whenever any axis reports moving
  This is new.
  All drivers are oblidged to do the right thing here,
  when they want to avoid fast polling in an error case.

Note:
I have seen controllers which are connected through TCP/IP
and they return crap (old responses in the outbuffer).
When the IOC is restarted and e.g. the firmware version must
be read from the controller.
After the startup phase the poll() will only report asynError
when there is a connection problem.